+++ lib/td_audit/notifications/dispatcher.ex
@@ -26,8 +26,14 @@ defmodule TdAudit.Notifications.Dispatcher do
 
   def send_email(email) do
     case Mailer.deliver_now(email, response: true) do
-      {:ok, _email, response} -> Logger.info("Email sent: Obtained #{response} from server")
-      {:error, error} -> Logger.error("Error: #{error}")
+      {:ok, response} ->
+        Logger.info("Email sent: Obtained #{inspect(response)} from server")
+
+      {:ok, _email, response} ->
+        Logger.info("Email sent: Obtained #{inspect(response)} from server")
+
+      {:error, error} ->
+        Logger.error("Error: #{error}")
     end
   end