+++ test/td_audit/subscriptions/subscription_test.exs
@@ -21,5 +21,47 @@ defmodule TdAudit.Subscriptions.SubscriptionTest do
 
       assert %{events: [{"can't be blank", [validation: :required]}]} = scope_errors
     end
+
+    test "includes resource_name and domain_id in scope Scope.changeset/2" do
+      %{
+        last_event_id: last_event_id,
+        periodicity: periodicity,
+        scope: %{
+          "resource_id" => resource_id,
+          "resource_name" => resource_name,
+          "resource_type" => resource_type,
+          "events" => events,
+          "domain_id" => domain_id
+        }
+      } =
+        params = %{
+          scope: %{
+            "resource_id" => 42,
+            "resource_name" => "bar",
+            "resource_type" => "data_structure",
+            "events" => ["foo"],
+            "domain_id" => 1
+          },
+          periodicity: "hourly",
+          last_event_id: 0
+        }
+
+      %{
+        valid?: true,
+        changes: %{
+          last_event_id: ^last_event_id,
+          periodicity: ^periodicity,
+          scope: %{
+            changes: %{
+              events: ^events,
+              resource_id: ^resource_id,
+              resource_name: ^resource_name,
+              resource_type: ^resource_type,
+              domain_id: ^domain_id
+            }
+          }
+        }
+      } = Subscription.changeset(params)
+    end
   end
 end