+++ lib/td_audit/audit/event.ex
@@ -50,6 +50,7 @@ defmodule TdAudit.Audit.Event do
   defp validate_user_and_resource(changeset) do
     case get_field(changeset, :event) do
       "login_attempt" -> changeset
+      "share_document" -> changeset
       _ -> validate_required(changeset, [:user_id, :resource_type, :resource_id])
     end
   end