+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## [Unreleased] 
+## [Unreleased]
 
 ### Added
 
 - [TD-3346] Search events Api
+
+## Fixed
+
+- [TD-3618] Exception sending email
 
 ## [4.17.0] 2021-04-05
 
@@ -37,7 +41,7 @@
 - [TD-3265] Url for concept resources in notifications
 - [TD-3245] Tested compatibility with PostgreSQL 9.6, 10.15, 11.10, 12.5 and
   13.1. CI pipeline changed to use `postgres:12.5-alpine`.
-  
+
 ### Removed
 
 - [TD-3171] Kubernetes jobs are no longer launched by `td-audit`