+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-3503] Share notification: Retrieve user emails by id
+- [TD-3753] Build using Elixir 1.12 and Erlang/OTP 24
 
 ## [4.20.0] 2021-05-17